+++ components/share/receive/src/main/java/com/flipperdevices/share/receive/models/FlipperKeyParseException.kt
@@ -1,0 +1,3 @@
+package com.flipperdevices.share.receive.models
+
+class FlipperKeyParseException : Exception()